Moving to `GatherModelPartUtility`
