+++ kratos/includes/fill_communicator.h
@@ -115,58 +115,6 @@ public:
     virtual void Execute();
 
     /**
-     * @brief Function to gather entities from other partitions
-     * @details This function is intended to gather entities from other partitions. The map indicates the partitions to gather and the vector the entities to gather from each partition. In the current serial case it does nothing.
-     * @note For the parallel implementation see ParallelFillCommunicator.
-     * @param rNodesToBring Nodes to gather from other partitions
-     * @param rElementsToBring Elements to gather from other partitions
-     * @param rConditionsToBring Conditions to gather from other partitions
-     * @param CallExecuteAfterBringingEntities Call Execute after gathering entities
-     */
-    virtual void GatherEntitiesFromOtherPartitions(
-        const std::map<int, std::vector<std::size_t>>& rNodesToBring,
-        const std::map<int, std::vector<std::size_t>>& rElementsToBring,
-        const std::map<int, std::vector<std::size_t>>& rConditionsToBring,
-        const bool CallExecuteAfterBringingEntities = true
-        );
-
-    /**
-     * @brief Function to gather nodes from other partitions
-     * @details This function is intended to gather nodes from other partitions. The map indicates the partitions to gather and the vector the nodes to gather from each partition. In the current serial case it does nothing.
-     * @note For the parallel implementation see ParallelFillCommunicator.
-     * @param rNodesToBring Nodes to gather from other partitions
-     * @param CallExecuteAfterBringingEntities Call Execute after gathering nodes
-     */
-    virtual void GatherNodesFromOtherPartitions(
-        const std::map<int, std::vector<std::size_t>>& rNodesToBring,
-        const bool CallExecuteAfterBringingEntities = true
-        );
-
-    /**
-     * @brief Function to gather elements from other partitions
-     * @details This function is intended to gather elements from other partitions. The map indicates the partitions to gather and the vector the elements to gather from each partition. In the current serial case it does nothing.
-     * @note For the parallel implementation see ParallelFillCommunicator.
-     * @param rElementsToBring Elements to gather from other partitions
-     * @param CallExecuteAfterBringingEntities Call Execute after gathering elements
-     */
-    virtual void GatherElementsFromOtherPartitions(
-        const std::map<int, std::vector<std::size_t>>& rElementsToBring,
-        const bool CallExecuteAfterBringingEntities = true
-        );
-
-    /**
-     * @brief Function to gather conditions from other partitions
-     * @details This function is intended to gather conditions from other partitions. The map indicates the partitions to gather and the vector the conditions to gather from each partition. In the current serial case it does nothing.
-     * @note For the parallel implementation see ParallelFillCommunicator.
-     * @param rConditionsToBring Conditions to gather from other partitions
-     * @param CallExecuteAfterBringingEntities Call Execute after gathering conditions
-     */
-    virtual void GatherConditionsFromOtherPartitions(
-        const std::map<int, std::vector<std::size_t>>& rConditionsToBring,
-        const bool CallExecuteAfterBringingEntities = true
-        );
-
-    /**
      * @brief Function to print DETAILED mesh information
      * WARNING: to be used for debugging only as many information are plotted
      */
