+++ kratos/mpi/utilities/parallel_fill_communicator.h
@@ -5,7 +5,7 @@
 //                   Multi-Physics
 //
 //  License:         BSD License
-//	                 Kratos default license: kratos/license.txt
+//                   Kratos default license: kratos/license.txt
 //
 //  Main authors:    Riccardo Rossi
 //
@@ -98,57 +98,6 @@ public:
      * @details This method is intended to perform the communicator filling
      */
     void Execute() override;
-
-    /**
-     * @brief Function to gather entities from other partitions
-     * @details This function is intended to gather entities from other partitions. The map indicates the partitions to gather and the vector the entities to gather from each partition. In the current serial case it does nothing.
-     * @param rNodesToBring Nodes to gather from other partitions
-     * @param rElementsToBring Elements to gather from other partitions
-     * @param rConditionsToBring Conditions to gather from other partitions
-     * @param CallExecuteAfterBringingEntities Call Execute after gathering entities
-     */
-    void GatherEntitiesFromOtherPartitions(
-        const std::map<int, std::vector<std::size_t>>& rNodesToBring,
-        const std::map<int, std::vector<std::size_t>>& rElementsToBring,
-        const std::map<int, std::vector<std::size_t>>& rConditionsToBring,
-        const bool CallExecuteAfterBringingEntities = true
-        ) override;
-
-    /**
-     * @brief Function to gather nodes from other partitions
-     * @details This function is intended to gather nodes from other partitions. The map indicates the partitions to gather and the vector the nodes to gather from each partition. In the current serial case it does nothing.
-     * @note For the parallel implementation see ParallelFillCommunicator.
-     * @param rNodesToBring Nodes to gather from other partitions
-     * @param CallExecuteAfterBringingEntities Call Execute after gathering nodes
-     */
-    void GatherNodesFromOtherPartitions(
-        const std::map<int, std::vector<std::size_t>>& rNodesToBring,
-        const bool CallExecuteAfterBringingEntities = true
-        ) override;
-
-    /**
-     * @brief Function to gather elements from other partitions
-     * @details This function is intended to gather elements from other partitions. The map indicates the partitions to gather and the vector the elements to gather from each partition. In the current serial case it does nothing.
-     * @note For the parallel implementation see ParallelFillCommunicator.
-     * @param rElementsToBring Elements to gather from other partitions
-     * @param CallExecuteAfterBringingEntities Call Execute after gathering elements
-     */
-    void GatherElementsFromOtherPartitions(
-        const std::map<int, std::vector<std::size_t>>& rElementsToBring,
-        const bool CallExecuteAfterBringingEntities = true
-        ) override;
-
-    /**
-     * @brief Function to gather conditions from other partitions
-     * @details This function is intended to gather conditions from other partitions. The map indicates the partitions to gather and the vector the conditions to gather from each partition. In the current serial case it does nothing.
-     * @note For the parallel implementation see ParallelFillCommunicator.
-     * @param rConditionsToBring Conditions to gather from other partitions
-     * @param CallExecuteAfterBringingEntities Call Execute after gathering conditions
-     */
-    void GatherConditionsFromOtherPartitions(
-        const std::map<int, std::vector<std::size_t>>& rConditionsToBring,
-        const bool CallExecuteAfterBringingEntities = true
-        ) override;
 
     /**
      * @brief Function to print mesh information of the provided model part
@@ -250,18 +199,6 @@ private:
     ///@}
     ///@name Private Operations
     ///@{
-
-    /**
-     * @brief Function to gather entities from other partitions
-     * @details This function is intended to gather entities from other partitions. The map indicates the partitions to gather and the vector the entities to gather from each partition. In the current serial case it does nothing.
-     * @param rModelPart Model part to gather entities from other partitions
-     * @param rEntitiesToBring Entities to gather from other partitions
-     */
-    template <class TObjectType>
-    void GatherEntityFromOtherPartitions(
-        ModelPart& rModelPart,
-        const std::map<int, std::vector<std::size_t>>& rEntitiesToBring
-        );
 
     ///@}
     ///@name Private  Access
