+++ kratos/python/add_other_utilities_to_python.cpp
@@ -684,10 +684,6 @@ void AddOtherUtilitiesToPython(pybind11::module &m)
         }) )
         .def(py::init<ModelPart&, const DataCommunicator& >() )
         .def("Execute", &FillCommunicator::Execute)
-        .def("GatherEntitiesFromOtherPartitions", &FillCommunicator::GatherEntitiesFromOtherPartitions)
-        .def("GatherNodesFromOtherPartitions", &FillCommunicator::GatherNodesFromOtherPartitions)
-        .def("GatherElementsFromOtherPartitions", &FillCommunicator::GatherElementsFromOtherPartitions)
-        .def("GatherConditionsFromOtherPartitions", &FillCommunicator::GatherConditionsFromOtherPartitions)
         .def("PrintDebugInfo", &FillCommunicator::PrintDebugInfo)
         .def("SetEchoLevel", &FillCommunicator::SetEchoLevel)
         .def("GetEchoLevel", &FillCommunicator::GetEchoLevel)
