+++ kratos/utilities/fill_communicator.cpp
@@ -39,36 +39,6 @@ void FillCommunicator::Execute()
     KRATOS_CATCH("");
 }
 
-void FillCommunicator::GatherEntitiesFromOtherPartitions(
-    const std::map<int, std::vector<std::size_t>>& rNodesToBring,
-    const std::map<int, std::vector<std::size_t>>& rElementsToBring,
-    const std::map<int, std::vector<std::size_t>>& rConditionsToBring,
-    const bool CallExecuteAfterBringingEntities
-    )
-{
-}
-
-void FillCommunicator::GatherNodesFromOtherPartitions(
-    const std::map<int, std::vector<std::size_t>>& rNodesToBring,
-    const bool CallExecuteAfterBringingEntities
-    )
-{
-}
-
-void FillCommunicator::GatherElementsFromOtherPartitions(
-    const std::map<int, std::vector<std::size_t>>& rElementsToBring,
-    const bool CallExecuteAfterBringingEntities
-    )
-{
-}
-
-void FillCommunicator::GatherConditionsFromOtherPartitions(
-    const std::map<int, std::vector<std::size_t>>& rConditionsToBring,
-    const bool CallExecuteAfterBringingEntities
-    )
-{
-}
-
 void FillCommunicator::PrintDebugInfo()
 {
     PrintModelPartDebugInfo(mrBaseModelPart);
